requirements.txt
Django==1.11.3
django-braces==1.11.0
django-extensions==1.8.1
django-tinymce==2.6.0
django-widget-tweaks==1.4.1
djangorestframework==3.6.3
gunicorn==19.7.1
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.3
pytz==2017.2
six==1.10.0
